set up integration test and remove unnecessary folders and files
