--- test/public/index.test.js
@@ -1,7 +1,0 @@
-/* eslint-disable eol-last */
-/* eslint-disable no-undef */
-const staticFiles = require('../../public/index');
-
-test('public tests run', () => {
-  expect(true).toBe(true);
-});
