--- test/server/index.test.js
@@ -1,7 +1,0 @@
-/* eslint-disable eol-last */
-/* eslint-disable no-undef */
-const server = require('../../server/index');
-
-test('server tests run', () => {
-  expect(true).toBe(true);
-});
